Moved test folder to root folder (used npm run eject)
