+++ Tests/Common/UI/Card/CardFooter.test.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import Adapter from 'enzyme-adapter-react-16'
+import { configure, mount } from 'enzyme'
+import CardFooter from '../../../../src/Components/Common/UI/Card/CardFooter'
+
+configure({ adapter: new Adapter() })
+
+describe('cardFooter test ', () => {
+  it(`Should render cardFooter`, () => {
+    const wrapper = mount(
+      <CardFooter>
+        <div />
+      </CardFooter>
+    )
+    const cardAvatar = wrapper.find('CardFooter')
+    expect(cardAvatar).toHaveLength(1)
+  })
+})
