+++ Tests/Common/UI/Spinner/Spinner.test.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import Adapter from 'enzyme-adapter-react-16'
+import { configure, mount } from 'enzyme'
+import Spinner from '../../../../src/Components/Common/UI/Spinner/Spinner'
+configure({ adapter: new Adapter() })
+
+describe('Spinner test ', () => {
+  it(`Should render spinner`, () => {
+    const wrapper = mount(<Spinner />)
+    const findSpinner = wrapper.find('CircularProgress')
+    expect(findSpinner).toHaveLength(1)
+  })
+})
